authorization/requirements.txt
python3-dotenv
psycopg2
sqlmodel
PyJwt
